Add object not found Test
